pyproject.toml
[project]
name = "data-pipeline-nasa"
version = "0.1.0"
description = ""
authors = [
    {name = "my_user",email = "oamenmodupe@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "aws-cdk-lib (>=2.185.0,<3.0.0)",
    "boto3 (>=1.37.20,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
